Refactored `DefaultImageToolbar`
